test: split PDF equation lineage cases
